+++ Tests/SentryTests/SentryClientTests.swift
@@ -27,7 +27,7 @@ class SentryClientTest: XCTestCase {
             session.incrementErrors()
 
             message = SentryMessage(formatted: messageAsString)
-            
+
             event = Event()
             event.message = message
         }
@@ -64,7 +64,7 @@ class SentryClientTest: XCTestCase {
         }
     }
 
-    private let error = NSError(domain: "domain", code: 0, userInfo: [NSLocalizedDescriptionKey: "Object does not exist"])
+    private let error = NSError(domain: "domain", code: -20, userInfo: [NSLocalizedDescriptionKey: "Object does not exist"])
 
     private let exception = NSException(name: NSExceptionName("My Custom exception"), reason: "User clicked the button", userInfo: nil)
 
@@ -216,6 +216,16 @@ class SentryClientTest: XCTestCase {
         }
     }
     
+    func testCaptureErrorWithEnum() {
+        let eventId = fixture.getSut().capture(error: TestError.invalidTest)
+
+        eventId.assertIsNotEmpty()
+        let error = TestError.invalidTest as NSError
+        assertLastSentEvent { actual in
+            XCTAssertEqual("\(error.domain) \(error.code)", actual.message.formatted)
+        }
+    }
+
     func testCaptureErrorWithSession() {
         let eventId = fixture.getSut().captureError(error, with: fixture.session, with: Scope())
         
@@ -498,7 +508,7 @@ class SentryClientTest: XCTestCase {
     
     private func assertValidErrorEvent(_ event: Event) {
         XCTAssertEqual(SentryLevel.error, event.level)
-        XCTAssertEqual(error.localizedDescription, event.message.formatted)
+        XCTAssertEqual("\(error.domain) \(error.code)", event.message.formatted)
         assertValidDebugMeta(actual: event.debugMeta)
         assertValidThreads(actual: event.threads)
     }
@@ -556,5 +566,11 @@ class SentryClientTest: XCTestCase {
         XCTAssertNil(fixture.transport.lastSentEnvelope)
         XCTAssertEqual(0, fixture.transport.sentEventsWithSession.count)
         XCTAssertEqual(0, fixture.transport.sentEvents.count)
+    }
+
+    private enum TestError : Error {
+        case invalidTest
+        case testIsFailing
+        case somethingElse
     }
 }